refactor: club entity package 이름 변경
