+++ src/main/kotlin/join/famila/club/infrastructure/Category.kt
@@ -1,4 +1,4 @@
-package join.famila.persistent.entity.club
+package join.famila.club.infrastructure
 
 import jakarta.persistence.Embeddable
 import jakarta.persistence.EnumType.STRING
