+++ src/main/kotlin/join/famila/club/infrastructure/Club.kt
@@ -1,4 +1,4 @@
-package join.famila.persistent.entity.club
+package join.famila.club.infrastructure
 
 import jakarta.persistence.CascadeType
 import jakarta.persistence.CollectionTable
