+++ src/main/kotlin/join/famila/club/infrastructure/Member.kt
@@ -1,4 +1,4 @@
-package join.famila.persistent.entity.club
+package join.famila.club.infrastructure
 
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
@@ -7,31 +7,22 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
-import jakarta.persistence.Table
 import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
 import java.time.LocalDateTime
 import java.time.LocalDateTime.now
 
 @Entity
-class Picture(
+class Member(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0,
+    val id: Long,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
     val club: Club,
 
-    val memberId: Long,
-
-    val link: String,
+    val userId: Long,
 
     @CreatedDate
     val createdAt: LocalDateTime = now(),
-
-    @LastModifiedDate
-    val updatedAt: LocalDateTime? = null,
-
-    val deletedAt: LocalDateTime? = null,
 )
