+++ src/main/kotlin/join/famila/club/infrastructure/Picture.kt
@@ -1,9 +1,9 @@
-package join.famila.persistent.entity.club
+package join.famila.club.infrastructure
 
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
 import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType.IDENTITY
+import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
@@ -14,9 +14,9 @@ import java.time.LocalDateTime
 import java.time.LocalDateTime.now
 
 @Entity
-class Post(
+class Picture(
     @Id
-    @GeneratedValue(strategy = IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0,
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -25,17 +25,13 @@ class Post(
 
     val memberId: Long,
 
-    var title: String,
-
-    var content: String,
-
-    var isNotification: Boolean,
+    val link: String,
 
     @CreatedDate
     val createdAt: LocalDateTime = now(),
 
     @LastModifiedDate
-    var updatedAt: LocalDateTime? = null,
+    val updatedAt: LocalDateTime? = null,
 
-    var deletedAt: LocalDateTime? = null,
+    val deletedAt: LocalDateTime? = null,
 )
