+++ src/main/kotlin/join/famila/club/infrastructure/Post.kt
@@ -1,28 +1,41 @@
-package join.famila.persistent.entity.club
+package join.famila.club.infrastructure
 
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
 import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
+import jakarta.persistence.GenerationType.IDENTITY
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
+import jakarta.persistence.Table
 import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedDate
 import java.time.LocalDateTime
 import java.time.LocalDateTime.now
 
 @Entity
-class Member(
+class Post(
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long,
+    @GeneratedValue(strategy = IDENTITY)
+    val id: Long = 0,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
     val club: Club,
 
-    val userId: Long,
+    val memberId: Long,
+
+    var title: String,
+
+    var content: String,
+
+    var isNotification: Boolean,
 
     @CreatedDate
     val createdAt: LocalDateTime = now(),
+
+    @LastModifiedDate
+    var updatedAt: LocalDateTime? = null,
+
+    var deletedAt: LocalDateTime? = null,
 )
