+++ src/main/kotlin/join/famila/club/infrastructure/Tag.kt
@@ -1,4 +1,4 @@
-package join.famila.persistent.entity.club
+package join.famila.club.infrastructure
 
 enum class Tag(description: String) {
     HEALTH("헬스"),
